tests: DifferentVendorTest moved into Reflection/Parser
