+++ packages/Reflection/tests/Parser/UseDifferentVendorTest.php
@@ -1,13 +1,13 @@
 <?php declare(strict_types=1);
 
-namespace ApiGen\Tests\Generator;
+namespace ApiGen\Reflection\Tests\Parser;
 
 use ApiGen\Reflection\Parser\Parser;
 use ApiGen\Tests\AbstractContainerAwareTestCase;
 
 final class UseDifferentVendorTest extends AbstractContainerAwareTestCase
 {
-    public function test()
+    public function test(): void
     {
         $parser = $this->container->get(Parser::class);
